requirements.txt
streamlit
pyrebase4
firebase_admin
google-cloud-firestore
requests
